feat: Extract status change logic into custom hook
- Moved the status change handling logic from the ApplicationsTable component to a new custom hook, useStatusChange, for improved separation of concerns and code organization.
- The new hook includes validation and state management for application status updates, enhancing maintainability and clarity in the ApplicationsTable component.
