+++ src/app/(frontend)/(aet-app)/components/ApplicationsTable/index.tsx
@@ -43,6 +43,7 @@ import { PaidAtConfirmDialog } from './PaidAtConfirmDialog'
 import { useOfficeChange } from './hooks/useOfficeChange'
 import { useTableScroll } from './hooks/useTableScroll'
 import { useDueAmountChange } from './hooks/useDueAmountChange'
+import { useStatusChange } from './hooks/useStatusChange'
 import { TableScrollButtons } from './TableScrollButtons'
 
 export function ApplicationsTable({ dataFilter }: { dataFilter: string }) {
@@ -106,115 +107,13 @@ export function ApplicationsTable({ dataFilter }: { dataFilter: string }) {
 
   const { handleDueAmountChange } = useDueAmountChange(applications, setApplications)
 
-  const handleStatusChange = async (id: string, newStatus: string) => {
-    try {
-      const application = applications.find((app) => app.id === id)
-
-      if (!application) {
-        throw new Error('Application not found')
-      }
-
-      const currentStatus = application.status
-      const paymentStatus = application.payment_status
-
-      // Validate status change based on the rules
-      if (
-        newStatus === 'completed' &&
-        !(currentStatus === 'processing' && paymentStatus === 'paid')
-      ) {
-        toast({
-          title: 'Operation not allowed',
-          description:
-            'Only applications with status "Processing" and payment status "Paid" can be completed.',
-          variant: 'destructive',
-        })
-        return
-      }
-
-      if (
-        newStatus === 'cancelled' &&
-        !(
-          (currentStatus === 'processing' && paymentStatus !== 'paid') ||
-          currentStatus === 'submitted'
-        )
-      ) {
-        toast({
-          title: 'Operation not allowed',
-          description:
-            'Only applications with status "Submitted" or "Processing" (if not paid) can be cancelled.',
-          variant: 'destructive',
-        })
-        return
-      }
-
-      // Allow changing from submitted to processing if payment status is paid
-      if (
-        newStatus === 'processing' &&
-        !(currentStatus === 'submitted' && paymentStatus === 'paid')
-      ) {
-        toast({
-          title: 'Operation not allowed',
-          description:
-            'Only applications with status "Submitted" and payment status "Paid" can be marked as processing.',
-          variant: 'destructive',
-        })
-        return
-      }
-
-      // Open confirmation dialog
-      setPendingStatusChange({
-        id,
-        status: newStatus,
-        currentStatus,
-      })
-      setStatusConfirmDialogOpen(true)
-    } catch (error) {
-      console.error('Error preparing status change:', error)
-      toast({
-        title: 'Operation failed',
-        description: 'Could not prepare the status change. Please try again.',
-        variant: 'destructive',
-      })
-    }
-  }
-
-  const confirmStatusChange = async () => {
-    if (!pendingStatusChange) return
-
-    try {
-      const { id, status } = pendingStatusChange
-
-      const { error } = await supabase.from('fce_applications').update({ status }).eq('id', id)
-
-      if (error) throw error
-
-      // Update local state
-      setApplications((apps) =>
-        apps.map((app) =>
-          app.id === id
-            ? {
-                ...app,
-                status: status as 'completed' | 'cancelled' | 'draft' | 'submitted' | 'processing',
-              }
-            : app
-        )
-      )
-
-      toast({
-        title: 'Status updated',
-        description: `Application status has been changed to ${status}.`,
-      })
-    } catch (error) {
-      console.error('Error updating status:', error)
-      toast({
-        title: 'Update failed',
-        description: 'Could not update the status. Please try again.',
-        variant: 'destructive',
-      })
-    } finally {
-      setPendingStatusChange(null)
-    }
-  }
+  const { handleStatusChange, confirmStatusChange } = useStatusChange({
+    applications,
+    setApplications,
+    setPendingStatusChange,
+    setStatusConfirmDialogOpen,
+    pendingStatusChange,
+  })
 
   const handlePaymentStatusChange = async (
     id: string,
